Try an egui UI
